+++ inst/doc/rbacon.R
@@ -56,15 +56,15 @@ require(rbacon)
 Bacon("MSB2K", ask=FALSE, coredir=tempdir(), suggest=FALSE)
 agedepth()
 
-## -----------------------------------------------------------------------------
+## ----fig.width=5, fig.height=4------------------------------------------------
 Bacon.hist(20)
 
-## -----------------------------------------------------------------------------
+## ----fig.width=5, fig.height=4------------------------------------------------
 a.d20 <- Bacon.Age.d(20)
 summary(a.d20)
 hist(a.d20)
 
-## -----------------------------------------------------------------------------
+## ----fig.width=5, fig.height=4------------------------------------------------
 a.d30 <- Bacon.Age.d(30)
 a.d20 <- Bacon.Age.d(20)
 summary(a.d30-a.d20)
@@ -75,6 +75,11 @@ acc.d20 <- accrate.depth(20)
 summary(acc.d20)
 acc.a4500 <- accrate.age(4500)
 summary(acc.a4500)
+
+## -----------------------------------------------------------------------------
+accrate.depth.summary(20)
+accrate.age.summary(4500)
+accrates.core()
 
 ## ----eval=FALSE---------------------------------------------------------------
 #  Bacon.hist(23.45)
@@ -110,4 +115,8 @@ summary(acc.a4500)
 #  hist(passed)
 #  mean(passed)
 #  quantile(passed, probs=c(.025, .975))
+
+## ----eval=FALSE---------------------------------------------------------------
+#  my_fancy_colours <- sample(colours(), nrow(info$dets))
+#  agedepth(height=50, dates.col=my_fancy_colours)